# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lettura API",
    "description": "Lettura is a wallabag-inspired read-it-later service. Endpoints below are stable; unannotated routes exist but their schema is not yet machine-readable.",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/fetch-jobs": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "list",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of fetch jobs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/fetch-jobs/retry-all-dead": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "retry_all_dead",
        "responses": {
          "200": {
            "description": "Dead jobs retried",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryAllResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/fetch-jobs/{id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Fetch job ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch job details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchJobRow"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "Fetch job not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Fetch job ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Fetch job deleted"
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/fetch-jobs/{id}/retry": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "retry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Fetch job ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch job retried",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchJobRow"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "Fetch job not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FetchJobRow": {
        "type": "object",
        "required": [
          "id",
          "entry_id",
          "user_id",
          "url",
          "status",
          "priority",
          "attempts",
          "max_attempts",
          "run_after",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "attempts": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "entry_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_error": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_error_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "leased_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "leased_until": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "max_attempts": {
            "type": "integer",
            "format": "int32"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "refetch_requested_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "run_after": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/FetchJobStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "url": {
            "type": "string"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "FetchJobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "failed",
          "dead"
        ]
      },
...
      "ListResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FetchJobRow"
            }
          }
        }
      },
...
      "RetryAllResponse": {
        "type": "object",
        "required": [
          "retried",
          "remaining_dead"
        ],
        "properties": {
          "remaining_dead": {
            "type": "integer",
            "format": "int64"
          },
          "retried": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}